# src/explore-education-statistics-api-docs/source/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Explore education statistics - Public Data API",
    "contact": {
      "name": "Explore education statistics",
      "url": "https://explore-education-statistics.service.gov.uk",
      "email": "explore.statistics@education.gov.uk"
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/",
      "description": "API server"
    }
  ],
  "paths": {
...
    "/v1/publications": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "List publications",
        "description": "Lists details about publications with data available for querying.",
        "operationId": "ListPublications",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "A search term to find matching publications.",
            "schema": {
              "minLength": 3,
              "type": "string",
              "description": "A search term to find matching publications.",
              "nullable": true,
              "example": "Pupil absence"
            },
            "example": "Pupil absence"
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page of results to fetch.",
            "schema": {
              "minimum": 1,
              "type": "integer",
              "description": "The page of results to fetch.",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The maximum number of results per page.",
            "schema": {
              "maximum": 40,
              "minimum": 1,
              "type": "integer",
              "description": "The maximum number of results per page.",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The paginated list of publications.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationPaginatedListViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemViewModel"
                }
              }
            }
          },
          "504": {
            "description": "Timeout"
          }
        }
      }
    },
    "/v1/publications/{publicationId}": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "Get a publication’s details",
        "description": "Get a specific publication's summary details.",
        "operationId": "GetPublication",
        "parameters": [
          {
            "name": "publicationId",
            "in": "path",
            "description": "The ID of the publication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested publication summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationSummaryViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsViewModel"
                }
              }
            }
          },
          "504": {
            "description": "Timeout"
          }
        }
      }
    },
    "/v1/publications/{publicationId}/data-sets": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "List a publication’s data sets",
        "description": "Lists summary details of all the data sets related to a publication.",
        "operationId": "ListPublicationDataSets",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The page of results to fetch.",
            "schema": {
              "minimum": 1,
              "type": "integer",
              "description": "The page of results to fetch.",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The maximum number of results per page.",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "description": "The maximum number of results per page.",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "publicationId",
            "in": "path",
            "description": "The ID of the publication.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The paginated list of data sets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSetPaginatedListViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemViewModel"
                }
              }
            }
          },
          "504": {
            "description": "Timeout"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataSetLatestVersionViewModel": {
        "required": [
          "file",
          "filters",
          "geographicLevels",
          "indicators",
          "published",
          "timePeriods",
          "totalResults",
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "The version number. Follows semantic versioning e.g. 2.0 (major), 1.1 (minor).",
            "example": "2.0"
          },
          "published": {
            "type": "string",
            "description": "When the version was published.",
            "format": "date-time",
            "example": "2024-03-01T09:30:00+00:00"
          },
          "totalResults": {
            "type": "integer",
            "description": "The total number of results available to query in the data set.",
            "format": "int64",
            "example": 1000000
          },
          "file": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DataSetVersionFileViewModel"
              }
            ],
            "description": "The file that this data set version is based on."
          },
          "timePeriods": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimePeriodRangeViewModel"
              }
            ],
            "description": "The time period range covered by the data set."
          },
          "geographicLevels": {
            "type": "array",
            "items": {
              "enum": [
                "English devolved area",
                "Local authority",
                "Local authority district",
                "Local enterprise partnership",
                "Local skills improvement plan area",
                "Institution",
                "Mayoral combined authority",
                "MAT",
                "National",
                "Opportunity area",
                "Parliamentary constituency",
                "Provider",
                "Regional",
                "RSC region",
                "School",
                "Sponsor",
                "Ward",
                "Planning area"
              ],
              "type": "string"
            },
            "description": "The geographic levels available in the data set.",
            "example": [
              "National",
              "Regional",
              "Local authority"
            ]
          },
          "filters": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The filters available in the data set.",
            "example": [
              "Characteristic",
              "School type"
            ]
          },
          "indicators": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The indicators available in the data set."
          }
        },
        "additionalProperties": false,
        "description": "Provides high-level information about the latest version of a data set."
      },
...
      "DataSetPaginatedListViewModel": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "paging": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PagingViewModel"
              }
            ],
            "description": "Provides metadata for use in pagination."
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataSetViewModel"
            },
            "description": "The list of results for this page."
          }
        },
        "additionalProperties": false,
        "description": "A paginated list of data sets."
      },
...
      "DataSetStatus": {
        "enum": [
          "Published",
          "Deprecated",
          "Withdrawn"
        ],
        "type": "string"
      },
...
      "DataSetVersionFileViewModel": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the file.",
            "format": "uuid",
            "example": "e0754872-3206-4918-aad4-029eaaae191f"
          }
        },
        "additionalProperties": false,
        "description": "Details about the file a data set version is based on."
      },
...
      "DataSetViewModel": {
        "required": [
          "id",
          "latestVersion",
          "status",
          "summary",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the data set.",
            "format": "uuid",
            "example": "7588c2d6-9e8a-4d84-8f19-6b8d52a01fbd"
          },
          "title": {
            "type": "string",
            "description": "The title of the data set.",
            "example": "Absence rates by geographic level"
          },
          "summary": {
            "type": "string",
            "description": "A summary of the data set’s contents.",
            "example": "Absence information for all enrolments in schools."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DataSetStatus"
              }
            ],
            "description": "The status of the data set. Can be one of the following:\r\n            \r\n- `Published` - the data set has been published and will receive updates\r\n- `Deprecated` - the data set is being discontinued and will no longer receive updates\r\n- `Withdrawn` - the data set has been withdrawn and can no longer be used",
            "example": "Published"
          },
          "supersedingDataSetId": {
            "type": "string",
            "description": "The ID of the data set that supersedes this data set (if it has been deprecated).",
            "format": "uuid",
            "nullable": true,
            "example": "2118a6df-4934-4a1f-ad2e-4589d2b9ccaf"
          },
          "latestVersion": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DataSetLatestVersionViewModel"
              }
            ],
            "description": "The latest published data set version."
          }
        },
        "additionalProperties": false,
        "description": "Provides high-level information about a data set."
      },
      "ErrorViewModel": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message.",
            "example": "Must be 50 characters or fewer."
          },
          "path": {
            "type": "string",
            "description": "The path to the property on the request that the error relates to.\r\nMay be omitted or empty if no specific property of the\r\nrequest relates to the error (it is a 'global' error).",
            "nullable": true,
            "example": "person.name"
          },
          "code": {
            "type": "string",
            "description": "The error's machine-readable code. Can be used for further\r\nprocessing of the error before presenting to users.\r\nMay be omitted if there is none.",
            "example": "MaximumLength"
          },
          "detail": {
            "description": "Additional detail about the error that can be used to provide\r\nmore context to users. May be omitted if there is none.",
            "nullable": true,
            "example": {
              "maxLength": 50
            }
          }
        },
        "additionalProperties": false,
        "description": "Describes an error that occurred with the request. This will typically\r\nneed to be rectified before the request can be fully processed."
      },
...
      "PagingViewModel": {
        "required": [
          "page",
          "pageSize",
          "totalPages",
          "totalResults"
        ],
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "The current page number.",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "The maximum number of results per page.",
            "format": "int32",
            "example": 20
          },
          "totalResults": {
            "type": "integer",
            "description": "The total number of results across all pages.",
            "format": "int32",
            "example": 60
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages.",
            "format": "int32",
            "readOnly": true,
            "example": 3
          }
        },
        "additionalProperties": false
      },
      "ProblemDetailsViewModel": {
        "required": [
          "status",
          "title",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "minLength": 1,
            "type": "string",
            "description": "A URI reference (RFC3986) that identifies the problem type. It should\r\nprovide human-readable documentation for the problem type.",
            "example": "https://tools.ietf.org/html/rfc9110#section-15.5.1"
          },
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "A short, human-readable summary of the problem type.",
            "example": "One or more validation errors occurred."
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for the problem.",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to the problem.",
            "nullable": true,
            "example": "More detail about the problem."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem.\r\nIt may or may not yield further information about the problem.",
            "nullable": true,
            "example": "https://example.com"
          }
        },
        "additionalProperties": { },
        "description": "Describes a problem that occurred whilst processing the request.\r\nThe format is based on RFC7807."
      },
...
      "PublicationPaginatedListViewModel": {
        "required": [
          "paging",
          "results"
        ],
        "type": "object",
        "properties": {
          "paging": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PagingViewModel"
              }
            ],
            "description": "Provides metadata for use in pagination."
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationSummaryViewModel"
            },
            "description": "The list of results for this page."
          }
        },
        "additionalProperties": false,
        "description": "A paginated list of publication summaries."
      },
      "PublicationSummaryViewModel": {
        "required": [
          "id",
          "lastPublished",
          "slug",
          "summary",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the publication.",
            "format": "uuid",
            "example": "d851c09e-7f5a-4750-9191-ed67ba5e8f8b"
          },
          "title": {
            "type": "string",
            "description": "The title of the publication.",
            "example": "Pupil absence in schools in England"
          },
          "slug": {
            "type": "string",
            "description": "The URL slug of the publication.",
            "example": "pupil-absence-in-schools-in-england"
          },
          "summary": {
            "type": "string",
            "description": "The summary of the publication.",
            "example": "Summary of the publication."
          },
          "lastPublished": {
            "type": "string",
            "description": "When the publication was last published.",
            "format": "date-time",
            "example": "2024-06-01T09:30:00+00:00"
          }
        },
        "additionalProperties": false,
        "description": "Provides summary information about a publication."
      },
...
      "TimePeriodRangeViewModel": {
        "required": [
          "end",
          "start"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "description": "The starting time period in human-readable format.",
            "example": "2024 January"
          },
          "end": {
            "type": "string",
            "description": "The ending time period in human-readable format.",
            "example": "2024 December"
          }
        },
        "additionalProperties": false,
        "description": "Describes a time period range in human-readable format."
      },
...
      "ValidationProblemViewModel": {
        "required": [
          "title",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "minLength": 1,
            "type": "string",
            "description": "A URI reference (RFC3986) that identifies the problem type. It should\r\nprovide human-readable documentation for the problem type.",
            "example": "https://tools.ietf.org/html/rfc9110#section-15.5.1"
          },
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "A short, human-readable summary of the problem type.",
            "example": "One or more validation errors occurred."
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to the problem.",
            "nullable": true,
            "example": "More detail about the problem."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem.\r\nIt may or may not yield further information about the problem.",
            "nullable": true,
            "example": "https://example.com"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for the problem.",
            "format": "int32",
            "example": 400
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorViewModel"
            },
            "description": "The errors relating to the validation problem."
          }
        },
        "additionalProperties": { },
        "description": "A validation problem that has occurred with the API request.\r\nThis type of response will be composed of one or more errors\r\nrelating to the request and its properties."
      },
...
    }
  }
}